+++ src/PanCAD/cad/freecad/__init__.py
@@ -17,5 +17,4 @@ import Part
 import PartDesign
 
 from PanCAD.cad.freecad.file import File
-from PanCAD.cad.freecad.to_part_file import to_freecad
-from PanCAD.cad.freecad.read_freecad import FreeCADFile
+from PanCAD.cad.freecad.filetypes import FreeCADFile
